requirements.txt
aiogram>=3.15.0
python-dotenv>=1.0.0
rembg>=2.0.50
onnxruntime>=1.16.0
Pillow>=10.0.0
numpy>=1.24.0
opencv-python-headless>=4.9.0.80
